The proposed code adds pthreads handling to the KokkosBlas2_gemv_perf_test.  I added these changes because Brian Kelley mentioned that he'd not done a "--threads" implementation yet, so I was trying to help him out with this test.

Additionally, explanatory comments were added, so that this test can be used as a model for design of future performance tests.

The test builds and runs.  Test output for a pthreads implementation is provided below.  

[ajpowel@kokkos-dev-2 blas2]$ ./KokkosBlas2_gemv_perf_test 
Running GEMV experiment (Serial)
Avg GEMV time: 0.145189 s.
Avg GEMV FLOP/s: 3.444e+08
[ajpowel@kokkos-dev-2 blas2]$ 
[ajpowel@kokkos-dev-2 blas2]$ 
[ajpowel@kokkos-dev-2 blas2]$ ./KokkosBlas2_gemv_perf_test --threads 20
Running GEMV experiment (Threads)
Avg GEMV time: 0.008392 s.
Avg GEMV FLOP/s: 5.958e+09
Running GEMV experiment (Serial)
Avg GEMV time: 0.171651 s.
Avg GEMV FLOP/s: 2.913e+08
[ajpowel@kokkos-dev-2 blas2]$ 
[ajpowel@kokkos-dev-2 blas2]$ ./KokkosBlas2_gemv_perf_test --threads 40
Running GEMV experiment (Threads)
Avg GEMV time: 0.006137 s.
Avg GEMV FLOP/s: 8.147e+09
Running GEMV experiment (Serial)
Avg GEMV time: 0.204992 s.
Avg GEMV FLOP/s: 2.439e+08


[ajpowel@kokkos-dev-2 blas2]$ ./KokkosBlas2_gemv_perf_test --openmp 20
ERROR: OpenMP requested, but not available.
ERROR : Process exiting while Kokkos::Threads is still initialized

[ajpowel@kokkos-dev-2 blas2]$ ./KokkosBlas2_gemv_perf_test --openmp 40
ERROR: OpenMP requested, but not available.
ERROR : Process exiting while Kokkos::Threads is still initialized

[ajpowel@kokkos-dev-2 blas2]$ ./KokkosBlas2_gemv_perf_test --cuda
Segmentation fault (core dumped)